+++ docs/source/release_notes/release-notes.rst
@@ -63,6 +63,10 @@ develop
          - Persisted locks table is now considered an Atomic Table and will not get KVS migrated
            (`Pull Request <https://github.com/palantir/atlasdb/pull/1610>`__)
 
+    *    - |improved|
+         - Reduced contention on PersistentTimestampService.getFreshTimestamps.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/1618>`__)
+
 .. <<<<------------------------------------------------------------------------------------------------------------->>>>
 
 =======